refactor: 💡 rename postcodeSuggestion to agencySelection
