+++ src/resources/scss/styles.scss
@@ -57,7 +57,7 @@
 @import '../../components/sessionsList/scss/sessionsList.styles';
 @import '../../components/sessionsListItem/scss/sessionsListItem.styles';
 @import '../../components/typingIndicator/scss/typingIndicator.styles';
-@import '../../components/postcodeSuggestion/scss/postcodeSuggestion.styles';
+@import '../../components/agencySelection/scss/agencySelection.styles';
 @import '../../components/login/scss/login.styles';
 @import '../../components/sessionHeader/scss/sessionHeader.styles';
 @import '../../components/sessionHeader/scss/sessionHeader.yellowTheme.styles';
